requirements.txt
daytona
streamlit
python-dotenv
pandas 
OpenAI